Fixes after code review: rewritten struct constructors and updated main.go
