--- go.mod
@@ -1,3 +1,0 @@
-module VladOdintsov
-
-go 1.24.3
