--- module1/describer.go
@@ -1,5 +1,0 @@
-package module1
-
-type Describer interface {
-	Describe() string
-}
