+++ module1/task1/go.mod
@@ -1,0 +1,3 @@
+module VladOdintsov/module1/task1
+
+go 1.24.3
