+++ module1/task1/utils/utils.go
@@ -1,0 +1,7 @@
+package utils
+
+import "strings"
+
+func FormatSlice(slice []string) string {
+	return strings.Join(slice, ",")
+}
